+++ SO/SO2/SO2/Request.cs
@@ -3,7 +3,7 @@ namespace SO2;
 public class Request {
     public int Cylinder { get; }
     public int ArrivalTime { get; }
-    public int WaitingTime { get; private set; }
+    public long WaitingTime { get; private set; }
 
     public Request(int cylinder, int arrivalTime) {
         Cylinder = cylinder;
@@ -11,6 +11,6 @@ public class Request {
     }
 
     public void AddWaitingTime(int time) {
-        WaitingTime += time;
+        WaitingTime += (long)time;
     }
 }
